+++ src/main/java/com/nhnacademy/illuwa/IlluwaApplication.java
@@ -2,9 +2,11 @@ package com.nhnacademy.illuwa;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
 @EnableScheduling
+@EnableFeignClients
 @SpringBootApplication
 public class IlluwaApplication {
