+++ src/main/java/com/nhnacademy/illuwa/domain/coupons/external/clientDto/BookDto.java
@@ -1,4 +1,4 @@
-package com.nhnacademy.illuwa.domain.coupons.external.book;
+package com.nhnacademy.illuwa.domain.coupons.external.clientDto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
