+++ src/test/resources/application-db.yml
@@ -14,8 +14,8 @@ spring:
         format_sql: true
         use_sql_comments: true
 
-book-api:
-  url: http://localhost:10304/
+#book-api:
+#  url: http://localhost:10304/
 
 user-api:
   url: http://localho:10303/
